adds a new directory `out_dicom_jpg` and updates `main.py` to include a new argument `--directory` and a new function `create_output_dir()` to create the directory.
the `create_output_dir()` function now takes an argument `output` which is used to create the directory. if no argument is passed, the default directory `out_jpg_files` is used. the `get_arguments()` function has been updated to include the `--directory` argument and the `output` argument now has a default value of `out_jpg_files`.

the `autocrop` class has been updated to loop through all the images in the `sample` directory, convert them to jpgs, crop them and save them in the `out_jpg_files` directory. a random uuid is generated for each image and used as the filename.

the `readme.md` file has been updated to include the correct paths for the input and output images.