implement a new version of the dicomcrop package, which will include two new functions, edges and crop.
the edges function takes an image as an input and returns a string describing the edges of the image. It uses the library AutoCrop to calculate the coordinates of the image and returns them as a tuple.

the crop function takes an image and an output directory as parameters and crops the image to a specified size. It uses the AutoCrop library to crop the image and save it to the output directory.

these two functions will make it easier for developers to crop images accurately and quickly, and will make the dicomcrop package more useful.