adds a new `_hash` function to the `cli.py` file in the `dicomcrop` package. this function is used to generate a token string, and it takes an optional `encrypted` parameter to determine whether the token should be encrypted or not.
the `_hash` function replaces the existing `token` function, and it uses the `uuid` library to generate a unique identifier. if `encrypted` is set to `true`, the function will return the `uuid` string directly. otherwise, it will call the `generate_token` function from the `lib` module to generate the token. the `secret` variable is also updated to use the `os.environ.get` function to get the secret from the environment.