adds a new parameter, `data`, to the `_hash` function in `dicomcrop/cli.py`. this parameter is a `dict` object that can be used to store additional data. the `_hash` function now takes two parameters, `data` and `encrypted`, and returns a token string.
the `_hash` function now also generates a unique identifier, `_id`, using the `uuid` module. this identifier is used to generate the token string if `data` is not provided. the `generate_token` function is used to generate the token string, and it takes the `patientid` from the `data` parameter if it is provided.

the `crop` function in `dicomcrop/cli.py` has also been updated to take an additional parameter, `encrypted`, which is used to determine whether the token string should be encrypted or not. the `_hash` function is now called with the `encrypted` parameter, and the `_objects` parameter is passed to the `_hash` function to store the `patientid` if it is provided.

finally, the `secret` variable in `dicomcrop/lib.py` has been updated to use the `str` function to ensure that the `secret` is a string.